+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-13 13:48:56
-// Commit: fix(core): prevent double tap duplicate action in core
+// Updated at: 2026-08-13 16:09:31
+// Commit: refactor(core): optimize data flow between domain and data layers
 
 class coreComponentState {
-  final String commitRef = '3575';
-  final String lastMessage = 'fix(core): prevent double tap duplicate action in core';
-  final DateTime timestamp = DateTime.parse('2026-08-13T13:48:56+05:00');
+  final String commitRef = '3579';
+  final String lastMessage = 'refactor(core): optimize data flow between domain and data layers';
+  final DateTime timestamp = DateTime.parse('2026-08-13T16:09:31+05:00');
 }